GH-1496: Polished
